Additional semantics for s-t problems
If passing a dict or a list of lists of targets now the function takes care to
select only the return values corresponding to those targets and those only, and
return an object of the same format of the targets parameter (i.e. either
ndarray, dict, or list of lists).